Rewrite ir interpreter.
This is supposed to get closer to being able to evaluate multi-level
IR stacks, like Project(Scan).

Added "execution blocks" which are built from the IR.
Execution blocks (XBs) are separate from ir::Blocks so that:
1. lifetimes don't pollute the IR code, as they could get confusing when applying transformations on it.
2. if we replace Executions Blocks with code generation, the IR stays the same and the XBs change or go away.

The use of StreamingIterators is introduced in place of plain Iterators for these reasons:
1) so project can build a local row to return, and allow it to be used by reference,
  but not need to retain all computed rows.
2) to limit lifetime of borrows from Scans to allow freeing/unlocking pages behind.)

There is quite a bit of lifetime specifiers in this.  I'm not at all sure I got them "right" but it compiles.
The idea is to allow delaying copies as far up the execution graphs as necessary (for a given row, but still limiting
the lifetime of borrows using StreamingIterators).  To fully realize this, I'll need to build a project block that
can skip rows without converting them.  The code doesn't do that right now.

I tried to use Box<dyn Trait> but could not get the lifetime of the borrows of the boxed values to work.
So I ended up with a enum approach rather than `dyn`.

`cargo fmt` was run